dashboard_requirements.txt
# Streamlit Dashboard Requirements
streamlit==1.28.0
plotly==5.17.0
pandas==2.0.3
google-cloud-bigquery==3.11.4
folium==0.15.0
streamlit-folium==0.15.0